Backend/requirements.txt
fastapi
uvicorn
scikit-learn
joblib
python-multipart
librosa
pillow
pydantic
